setup test photo

+++ KatieComedy.App/Photos/PhotoService.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Hosting;
 using SixLabors.ImageSharp;
@@ -9,6 +10,7 @@ namespace KatieComedy.App.Photos;
 public class PhotoService(
     ApplicationDbContext dbContext,
     IWebHostEnvironment env,
+    IHttpContextAccessor httpContextAccessor,
     IOptions<PhotoOptions> options)
 {
     private string PhotoDirectoryPath => Path.Combine(env.WebRootPath, "photos");
@@ -16,7 +18,10 @@ public class PhotoService(
 
     public async Task<IReadOnlyList<Photo>> Get(CancellationToken cancel)
     {
-        var photos = await dbContext.Photos.ToListAsync(cancel);
+        var photos = await dbContext.Photos
+            .Include(x => x.Thumbnail)
+            .Where(x => x.Type == PhotoType.Photo)
+            .ToListAsync(cancel);
 
         return photos
             .OrderByDescending(x => x.Date.HasValue)
@@ -25,9 +30,29 @@ public class PhotoService(
             .Select(x => new Photo
             {
                 Id = x.Id,
-                Url = string.Empty
+                Url = GetUrl(x.Filename)!,
+                ThumbnailUrl = GetUrl(x.Thumbnail?.Filename)
             })
             .ToList();
+    }
+
+    public async Task<Photo> Get(int id)
+    {
+        var photo = await dbContext.Photos
+            .Include(x => x.Thumbnail)
+            .FirstOrDefaultAsync(x => x.Id == id);
+
+        if (photo is null)
+        {
+            throw new Exception("Photo not found.");
+        }
+
+        return new Photo
+        {
+            Id = photo.Id,
+            Url = GetUrl(photo.Filename)!,
+            ThumbnailUrl = GetUrl(photo.Thumbnail?.Filename)
+        };
     }
 
     public async Task Upload(PhotoUpload upload, CancellationToken cancel)
@@ -45,17 +70,39 @@ public class PhotoService(
         var thumbnailFilename = Guid.NewGuid().ToString() + ".jpg";
         var thumbnailFilepath = Path.Combine(PhotoDirectoryPath, thumbnailFilename);
         await image.SaveAsJpegAsync(thumbnailFilepath, cancel);
+
+        dbContext.Photos.Add(new Database.Photo
+        {
+            Type = PhotoType.Photo,
+            Filename = photoFilename,
+            Date = upload.Date,
+            Caption = upload.Caption,
+            Thumbnail = new Database.Photo
+            {
+                Type = PhotoType.Thumbnail,
+                Filename = thumbnailFilename
+            }
+        });
+
+        await dbContext.SaveChangesAsync(cancel);
     }
 
     public async Task Delete(int id)
     {
-        var photo = await dbContext.Photos.FindAsync(id) ?? throw new FileNotFoundException();
-        var filepath = Path.Combine(PhotoDirectoryPath, photo.Filename);
-        File.Delete(filepath);
-        dbContext.Photos.Remove(photo);
-        await dbContext.SaveChangesAsync();
+        var photo = await dbContext.Photos
+            .Include(x => x.Thumbnail)
+            .FirstOrDefaultAsync(x => x.Id == id) ?? throw new FileNotFoundException();
 
-        // delete thumbnail too
+        dbContext.Photos.Remove(photo);
+        File.Delete(Path.Combine(PhotoDirectoryPath, photo.Filename));
+
+        if (photo.Thumbnail is not null)
+        {
+            dbContext.Photos.Remove(photo.Thumbnail);
+            File.Delete(Path.Combine(PhotoDirectoryPath, photo.Thumbnail.Filename));
+        }
+
+        await dbContext.SaveChangesAsync();
     }
 
     public async Task Update(int id)
@@ -73,6 +120,11 @@ public class PhotoService(
 
         foreach (var file in dirInfo.GetFiles())
         {
+            if (file.Name == _options.TestPhotoFilename)
+            {
+                return;
+            }
+
             file.Delete();
         }
 
@@ -80,5 +132,34 @@ public class PhotoService(
         {
             dir.Delete(true);
         }
+    }
+
+    public async Task InitializeTestPhoto()
+    {
+        if (env.IsProduction())
+        {
+            return;
+        }
+
+        var bytes = File.ReadAllBytes(Path.Join(PhotoDirectoryPath, _options.TestPhotoFilename));
+
+        await Upload(new PhotoUpload
+        {
+            Data = bytes,
+            FileExtension = ".jpg",
+            Caption = "This a test photo",
+            Date = DateOnly.FromDateTime(DateTime.Now)
+        }, default);
+    }
+
+    private string? GetUrl(string? filename)
+    {
+        if (string.IsNullOrEmpty(filename))
+        {
+            return null;
+        }
+
+        var httpContext = httpContextAccessor.HttpContext!;
+        return $"{httpContext.Request.Scheme}://{httpContext.Request.Host.ToUriComponent()}/photos/{filename}";
     }
 }
